Added files, they will change
Ive have uploaded this so that we can unify and make sure it is functional so far, then I will go about adding traits and sub abilities. How are we coding this in the UI? how are we doing status effects like light sensitivity ? Other things that would effect how rolls and results are calculated?